requirements.txt
python-telegram-bot[job-queue]==22.6
openai>=1.30
anthropic>=0.40
python-dotenv>=1.0
